qlf_k4n8: fix arch to allow physical .input and .output blif
This commit fixes an issue in the physical description of the IO
pb_types, for which, if a block that is not a top-level input/output has
an output with no inputs, it is considered by VPR as a constant
generator.

Given that the IO physical mode has only two possible settings (0 ==
output mode, 1 == input mode), the FASM feature is output only in case
the input mode is selected. This will enable us to leave unaltered the
top level IOs and let VPR correctly detect clock nets.

The repacker needs adjustments, and this commit also enhances the
packing rules to deal with the changes in the physical description of
the IO pb_type.

Signed-off-by: Alessandro Comodi <acomodi@antmicro.com>